# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AITeachMe",
    "description": "本地优先的 AI 助教后端服务。",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/subjects/{subject}/files/upload": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload files and start parsing immediately",
        "operationId": "upload_files_api_v1_subjects__subject__files_upload_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files_api_v1_subjects__subject__files_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_FilesUploadData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "上传文件超出大小限制。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/files": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Get all subject files with full data",
        "operationId": "list_files_api_api_v1_subjects__subject__files_get",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_FilesData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/files/delete": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Delete files",
        "operationId": "delete_files_api_api_v1_subjects__subject__files_delete_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_FileDeleteData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "当前资源状态不允许执行该操作。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_FileDeleteData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileDeleteData"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[FileDeleteData]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
      "ApiResponse_FilesData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilesData"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[FilesData]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
      "ApiResponse_FilesUploadData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilesUploadData"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[FilesUploadData]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
...
      "Body_upload_files_api_v1_subjects__subject__files_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_files_api_v1_subjects__subject__files_upload_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "type": "null",
            "title": "Data",
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ErrorResponse",
        "description": "统一错误响应结构。",
        "example": {
          "code": 404,
          "message": "资源不存在"
        }
      },
...
      "FileAssetItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Asset filename."
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Runtime static URL for this asset."
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type",
            "description": "Asset mime type."
          },
          "asset_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Kind",
            "description": "Asset kind."
          },
          "page_num": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Num",
            "description": "Source page number."
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width",
            "description": "Asset width."
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height",
            "description": "Asset height."
          },
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text",
            "description": "OCR text extracted from this asset."
          }
        },
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "FileAssetItem",
        "description": "Public asset descriptor for one extracted file asset."
      },
      "FileDeleteData": {
        "properties": {
          "deleted_file_uids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deleted File Uids",
            "description": "Deleted public file UIDs."
          }
        },
        "type": "object",
        "title": "FileDeleteData",
        "description": "Delete result."
      },
      "FileDeleteRequest": {
        "properties": {
          "file_uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Uid",
            "description": "Single public file UID."
          },
          "file_uids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Uids",
            "description": "Multiple public file UIDs."
          }
        },
        "type": "object",
        "title": "FileDeleteRequest",
        "description": "Delete request."
      },
      "FileRecord": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid",
            "description": "Stable public file UID."
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Filename."
          },
          "filetype": {
            "type": "string",
            "title": "Filetype",
            "description": "File extension."
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatusValue",
            "description": "Task status."
          },
          "ingest_status": {
            "type": "string",
            "title": "Ingest Status",
            "description": "Ingest workflow status."
          },
          "markdown_ready": {
            "type": "boolean",
            "title": "Markdown Ready",
            "description": "Whether markdown is ready."
          },
          "asset_ready": {
            "type": "boolean",
            "title": "Asset Ready",
            "description": "Whether asset directory is ready."
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Failure reason."
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes",
            "description": "File size in bytes."
          },
          "detected_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language",
            "description": "Detected language."
          },
          "estimated_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Pages",
            "description": "Estimated pages."
          },
          "image_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Count",
            "description": "Extracted image count."
          },
          "parser_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Used",
            "description": "Parser used."
          },
          "markdown_content": {
            "type": "string",
            "title": "Markdown Content",
            "description": "Parsed markdown content.",
            "default": ""
          },
          "asset_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Base Url",
            "description": "Runtime static base URL for this file's assets."
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/FileAssetItem"
            },
            "type": "array",
            "title": "Assets",
            "description": "Extracted assets."
          },
          "classification_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Json",
            "description": "Classification result JSON."
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score",
            "description": "Parse quality score."
          },
          "digest_current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest Current Step",
            "description": "Current digest step."
          },
          "parse_metadata_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Metadata Json",
            "description": "Parse metadata JSON."
          },
          "latest_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Latest Updated At",
            "description": "Last updated time."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Created time."
          }
        },
        "type": "object",
        "required": [
          "uid",
          "filename",
          "filetype",
          "status",
          "ingest_status",
          "markdown_ready",
          "asset_ready",
          "latest_updated_at",
          "created_at"
        ],
        "title": "FileRecord",
        "description": "Unified file record for list, preview, and upload responses."
      },
      "FilesData": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject slug."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total file count."
          },
          "ready_count": {
            "type": "integer",
            "title": "Ready Count",
            "description": "Count of markdown-ready files."
          },
          "processing_count": {
            "type": "integer",
            "title": "Processing Count",
            "description": "Count of files still processing."
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count",
            "description": "Count of failed files."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FileRecord"
            },
            "type": "array",
            "title": "Items",
            "description": "Full file records."
          }
        },
        "type": "object",
        "required": [
          "subject",
          "total",
          "ready_count",
          "processing_count",
          "failed_count"
        ],
        "title": "FilesData",
        "description": "Aggregated subject files response."
      },
      "FilesUploadData": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject slug."
          },
          "filenames": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filenames",
            "description": "Uploaded filenames."
          },
          "uploaded_items": {
            "items": {
              "$ref": "#/components/schemas/FileRecord"
            },
            "type": "array",
            "title": "Uploaded Items",
            "description": "Uploaded file records."
          },
          "started_parse_count": {
            "type": "integer",
            "title": "Started Parse Count",
            "description": "Count of auto-started parse files.",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "subject",
          "filenames"
        ],
        "title": "FilesUploadData",
        "description": "Upload response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskStatusValue": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "TaskStatusValue",
        "description": "任务状态。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}